Block_Chain/requirements.txt
# Sentiment Oracle Backend Dependencies
# Core ML/NLP Libraries
torch>=2.0.0
transformers>=4.30.0
nltk>=3.8.1
textblob>=0.17.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# API & Web
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Twitter/X API (optional - requires API keys)
tweepy>=4.14.0

# Utilities
python-dotenv>=1.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Web3
web3>=6.11.0
eth-account>=0.9.0
